browser/__init__.py
"""
waf.browser -- Browser automation toolkit.
"""

from waf.browser.stealth_driver import StealthDriver
from waf.browser.fingerprint_evasion import FingerprintEvasion
from waf.browser.human_behavior import HumanBehavior, HoneypotDetector

__all__ = ["StealthDriver", "FingerprintEvasion", "HumanBehavior", "HoneypotDetector"]
