requirements.txt
# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.1
fake-useragent>=1.4.0

# Human Behavior Simulation
numpy>=1.24.0

# IP Rotation
requests>=2.31.0
PySocks>=1.7.1

# LLM Providers
google-generativeai>=0.3.0
openai>=1.6.0
anthropic>=0.8.0

# Form Reconnaissance
beautifulsoup4>=4.12.0
